[U] First step towards bigger things - state manager moved to chunk
