--- VoxeUnity/Assets/Voxelmetric/Code/Core/StateManager.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: d6def0571571aa64f8e81a0acce2683c
-folderAsset: yes
-timeCreated: 1466326100
-licenseType: Pro
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
