--- VoxeUnity/Assets/Voxelmetric/Code/Core/StateManager/ChunkStateManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 15b34a39aa96b3f498c3839bda4d0a0e
-timeCreated: 1466335133
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
