--- VoxeUnity/Assets/Voxelmetric/Code/Core/StateManager/ChunkStateManagerClient.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 21f702cd77388444da452379d7350098
-timeCreated: 1466326101
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
