+++ VoxeUnity/Assets/Voxelmetric/Code/Utilities/ChunkLoaders/LoadChunksSimple.cs
@@ -5,8 +5,8 @@ using UnityEngine.Profiling;
 using Voxelmetric.Code.Common;
 using Voxelmetric.Code.Common.Math;
 using Voxelmetric.Code.Core;
-using Voxelmetric.Code.Core.StateManager;
 using Voxelmetric.Code.Data_types;
+using Chunk = Voxelmetric.Code.Core.Chunk;
 using Vector3Int = Voxelmetric.Code.Data_types.Vector3Int;
 
 namespace Voxelmetric.Code.Utilities.ChunkLoaders
@@ -125,9 +125,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                 Chunk chunk = world.chunks.Get(ref chunkPos);
                 if (chunk==null)
                     return;
-
-                ChunkStateManagerClient stateManager = chunk.stateManager;
-
+                
                 int xd = (m_viewerPos.x-chunk.Pos.x)/Env.ChunkSize;
                 int yd = (m_viewerPos.y-chunk.Pos.y)/Env.ChunkSize;
                 int zd = (m_viewerPos.z-chunk.Pos.z)/Env.ChunkSize;
@@ -138,10 +136,10 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                 // Update visibility information
                 bool isVisible = Planes.TestPlanesAABB(m_cameraPlanes, ref chunk.WorldBounds);
 
-                stateManager.Visible = isVisible &&
-                                        xDist<=HorizontalChunkLoadRadius*HorizontalChunkLoadRadius &&
-                                        yDist<=VerticalChunkLoadRadius*VerticalChunkLoadRadius;
-                stateManager.PossiblyVisible = isVisible || FullLoadOnStartUp;
+                chunk.Visible = isVisible &&
+                    xDist<=HorizontalChunkLoadRadius*HorizontalChunkLoadRadius &&
+                    yDist<=VerticalChunkLoadRadius*VerticalChunkLoadRadius;
+                chunk.PossiblyVisible = isVisible || FullLoadOnStartUp;
 
                 return;
             }
@@ -182,11 +180,9 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                             if (chunk==null)
                                 continue;
 
-                            ChunkStateManagerClient stateManager = chunk.stateManager;
-                            
                             // Update visibility information
-                            stateManager.PossiblyVisible = FullLoadOnStartUp;
-                            stateManager.Visible = false;
+                            chunk.PossiblyVisible = FullLoadOnStartUp;
+                            chunk.Visible = false;
                         }
                     }
                 }
@@ -212,9 +208,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                             Chunk chunk = world.chunks.Get(ref chunkPos);
                             if (chunk==null)
                                 continue;
-
-                            ChunkStateManagerClient stateManager = chunk.stateManager;
-
+                            
                             int xd = (m_viewerPos.x-chunk.Pos.x)/Env.ChunkSize;
                             int yd = (m_viewerPos.y-chunk.Pos.y)/Env.ChunkSize;
                             int zd = (m_viewerPos.z-chunk.Pos.z)/Env.ChunkSize;
@@ -223,9 +217,10 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                             int yDist = yd*yd;
 
                             // Update visibility information
-                            stateManager.Visible = xDist<=HorizontalChunkLoadRadius*HorizontalChunkLoadRadius &&
-                                                   yDist<=VerticalChunkLoadRadius*VerticalChunkLoadRadius;
-                            stateManager.PossiblyVisible = true;
+                            chunk.Visible =
+                                xDist<=HorizontalChunkLoadRadius*HorizontalChunkLoadRadius &&
+                                yDist<=VerticalChunkLoadRadius*VerticalChunkLoadRadius;
+                            chunk.PossiblyVisible = true;
                         }
                     }
                 }
@@ -359,9 +354,8 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
 
                         if (FullLoadOnStartUp)
                         {
-                            ChunkStateManagerClient stateManager = chunk.stateManager;
-                            stateManager.PossiblyVisible = true;
-                            stateManager.Visible = false;
+                            chunk.PossiblyVisible = true;
+                            chunk.Visible = false;
                         }
 
                         m_updateRequests.Add(chunk);
@@ -385,7 +379,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                 ProcessChunk(chunk);
 
                 // Update the chunk if possible
-                if (chunk.stateManager.CanUpdate)
+                if (chunk.CanUpdate)
                 {
                     chunk.UpdateState();
                     chunk.UpdateRenderGeometry();
@@ -393,8 +387,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                 }
 
                 // Automatically collect chunks which are ready to be removed from the world
-                ChunkStateManagerClient stateManager = chunk.stateManager;
-                if (stateManager.IsStateCompleted(ChunkState.Remove))
+                if (chunk.IsStateCompleted(ChunkState.Remove))
                 {
                     // Remove the chunk from our provider and unregister it from chunk storage
                     world.chunks.Remove(chunk);
@@ -432,9 +425,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
         public void ProcessChunk(Chunk chunk)
         {
             Profiler.BeginSample("ProcessChunk");
-
-            ChunkStateManagerClient stateManager = chunk.stateManager;
-
+            
             int xd = (m_viewerPos.x-chunk.Pos.x)/Env.ChunkSize;
             int yd = (m_viewerPos.y-chunk.Pos.y)/Env.ChunkSize;
             int zd = (m_viewerPos.z-chunk.Pos.z)/Env.ChunkSize;
@@ -447,7 +438,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
             // Remove the chunk if it is too far away
             if (xDist>hRadius*hRadius || yDist>vRadius*vRadius)
             {
-                stateManager.SetStatePending(ChunkState.Remove);
+                chunk.SetStatePending(ChunkState.Remove);
             }
             else
             {
@@ -457,9 +448,10 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                 if (!UseFrustumCulling)
                 {
                     // Update visibility information
-                    stateManager.Visible = xDist<=HorizontalChunkLoadRadius*HorizontalChunkLoadRadius &&
-                                           yDist<=VerticalChunkLoadRadius*VerticalChunkLoadRadius;
-                    stateManager.PossiblyVisible = true;
+                    chunk.Visible =
+                        xDist<=HorizontalChunkLoadRadius*HorizontalChunkLoadRadius &&
+                        yDist<=VerticalChunkLoadRadius*VerticalChunkLoadRadius;
+                    chunk.PossiblyVisible = true;
                 }
             }
 
@@ -577,8 +569,7 @@ namespace Voxelmetric.Code.Utilities.ChunkLoaders
                         }
 
                         // Show generated chunks
-                        ChunkStateManagerClient stateManager = chunk.stateManager;
-                        if (stateManager.IsStateCompleted(ChunkState.Generate))
+                        if (chunk.IsStateCompleted(ChunkState.Generate))
                         {
                             Gizmos.color = Color.magenta;
                             Gizmos.DrawWireCube(
